opinions

+++ go-api/internal/infrastracture/postgres/opinion.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 
 	"github.com/pttrulez/investor-go/internal/entity"
+	"github.com/pttrulez/investor-go/internal/infrastracture/database"
 )
 
 type OpinionPostgres struct {
@@ -16,33 +17,47 @@ func NewOpinionPostgres(db *sql.DB) *OpinionPostgres {
 	return &OpinionPostgres{db: db}
 }
 
-func (pg *OpinionPostgres) Delete(ctx context.Context, id int) error {
+func (pg *OpinionPostgres) Delete(ctx context.Context, id int, userID int) error {
 	const op = "OpinionPostgres.Delete"
 
-	queryString := "DELETE FROM opinions where id = $1;"
-	_, err := pg.db.ExecContext(ctx, queryString, id)
+	queryString := "DELETE FROM opinions where id = $1 and user_id = $2;"
+	result, err := pg.db.ExecContext(ctx, queryString, id, userID)
 	if err != nil {
 		return fmt.Errorf("%s: %w", op, err)
 	}
+
+	rowsAffected, err := result.RowsAffected()
+	if err != nil {
+		return fmt.Errorf("%s: %w", op, err)
+	}
+	if rowsAffected == 0 {
+		return database.ErrNotFound
+	}
+
 	return nil
 }
 
-func (pg *OpinionPostgres) Insert(ctx context.Context, o *entity.Opinion) error {
+func (pg *OpinionPostgres) Insert(ctx context.Context, o entity.Opinion) (entity.Opinion, error) {
 	const op = "OpinionPostgres.Insert"
 
 	queryString := `INSERT INTO opinions (date, exchange, expert_id, text, security_id,  
-    security_type, source_link, target_price, type, user_id) 
-    VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10);`
+    security_type, source_link, target_price, type, user_id) VALUES ($1, $2, $3, $4, $5,
+		$6, $7, $8, $9, $10) RETURNING id, date, exchange, expert_id, text, security_id,
+    security_type, source_link, target_price, type;`
 
-	_, err := pg.db.ExecContext(ctx, queryString, o.Date, o.Exchange, o.ExpertID, o.Text, o.SecurityID,
-		o.SecurityType, o.SourceLink, o.TargetPrice, o.Type, o.UserID)
+	var r entity.Opinion
+	err := pg.db.QueryRowContext(ctx, queryString, o.Date, o.Exchange, o.ExpertID, o.Text, o.SecurityID,
+		o.SecurityType, o.SourceLink, o.TargetPrice, o.Type, o.UserID).
+		Scan(&r.ID, &r.Date, &r.Exchange, &r.ExpertID, &r.Text, &r.SecurityID, &r.SecurityType,
+			&r.SourceLink, &r.TargetPrice, &r.Type)
 	if err != nil {
-		return fmt.Errorf("%s: %w", op, err)
+		return entity.Opinion{}, fmt.Errorf("%s: %w", op, err)
 	}
-	return nil
+
+	return r, nil
 }
 
-func (pg *OpinionPostgres) Update(ctx context.Context, o *entity.Opinion) error {
+func (pg *OpinionPostgres) Update(ctx context.Context, o entity.Opinion) error {
 	const op = "OpinionPostgres.Update"
 
 	queryString := `UPDATE opinions SET date = $1, exchange = $2, expert_id = $3,
@@ -57,22 +72,48 @@ func (pg *OpinionPostgres) Update(ctx context.Context, o *entity.Opinion) error 
 	return nil
 }
 
-func (pg *OpinionPostgres) GetListByUserID(ctx context.Context, userID int) ([]*entity.Opinion, error) {
+func (pg *OpinionPostgres) GetOpinionsList(ctx context.Context, f entity.OpinionFilters,
+	userID int) ([]entity.Opinion, error) {
 	const op = "OpinionPostgres.GetListByUserId"
 
-	queryString := `SELECT * FROM opinions WHERE user_id = $1;`
-	rows, err := pg.db.QueryContext(ctx, queryString, userID)
+	queryString := `SELECT id, date, exchange, expert_id, text, security_id, security_type,
+	 source_link, target_price, type FROM opinions WHERE user_id = $1`
+	args := []interface{}{userID}
+	count := 2
+
+	if f.ExpertID != nil {
+		queryString += fmt.Sprintf(" AND expert_id = $%d", count)
+		args = append(args, *f.ExpertID)
+		count++
+	}
+	if f.Exchange != nil {
+		queryString += fmt.Sprintf(" AND exchange = $%d", count)
+		args = append(args, *f.Exchange)
+		count++
+	}
+	if f.SecurityID != nil {
+		queryString += fmt.Sprintf(" AND security_id = $%d", count)
+		args = append(args, *f.SecurityID)
+		count++
+	}
+	if f.SecurityType != nil {
+		queryString += fmt.Sprintf(" AND security_type = $%d", count)
+		args = append(args, *f.SecurityType)
+	}
+	queryString += ";"
+
+	rows, err := pg.db.QueryContext(ctx, queryString, args...)
 	if err != nil {
 		return nil, fmt.Errorf("%s: %w", op, err)
 	}
 	defer rows.Close()
 
-	var opinions []*entity.Opinion
+	var opinions = make([]entity.Opinion, 0)
 
 	for rows.Next() {
-		o := new(entity.Opinion)
-		err = rows.Scan(&o.ID, &o.Date, &o.Exchange, &o.ExpertID, &o.SecurityID,
-			&o.SecurityType, &o.SourceLink, &o.TargetPrice, &o.Type, &o.UserID)
+		var o entity.Opinion
+		err = rows.Scan(&o.ID, &o.Date.Time, &o.Exchange, &o.ExpertID, &o.Text, &o.SecurityID,
+			&o.SecurityType, &o.SourceLink, &o.TargetPrice, &o.Type)
 		if err != nil {
 			return nil, fmt.Errorf("%s: %w", op, err)
 		}
